Attempt to center the empty state, move type defs around
