+++ src/__types__/index.ts
@@ -19,3 +19,22 @@ export type Icon = {
 export type MatIconList = {
     [key: string]: (props: SvgIconProps) => JSX.Element;
 };
+
+export type Status = 'backlog' | 'in-progress' | 'finished';
+export type Quarter = 'Q1' | 'Q2' | 'Q3' | 'Q4';
+
+export type RoadmapItem = {
+    name: string;
+    description: string;
+    year: string | number;
+    quarter: Quarter;
+    status: Status;
+    author?: string;
+    applies?: FrameworkFilter[];
+};
+export type RoadmapBucket = {
+    name: string;
+    description: string;
+    applies?: FrameworkFilter[];
+    items: RoadmapItem[];
+};
